fix(social/publish-due): extract claimDueDrafts to its own module
Next.js route files can only export the Route export fields
(GET/POST/runtime/etc). Exporting claimDueDrafts from the route file
broke the build with "claimDueDrafts is not a valid Route export field".

Move the function to lib/social/publishing/claim-due-drafts.ts; route +
test both import from there. Also makes the function more reusable for
follow-up uses (stale-claim reaper, manual republish endpoint).
